requirements.txt
customtkinter>=5.2.2
Pillow>=10.0.0
psutil>=5.9.0
screeninfo>=0.8.1
requests>=2.31.0
